feat: refacto metrics
