+++ src/github/metrics/active-contribution-weeks/active-contribution-weeks.metric.spec.ts
@@ -1,5 +1,5 @@
+import response from '@/github/metrics/query-response.example.json';
 import { ActiveContributionWeeksMetric } from './active-contribution-weeks.metric';
-import response from '../query-response.example.json';
 
 describe('Active Contribution Weeks Metric', () => {
   const metric = new ActiveContributionWeeksMetric();
