+++ src/github/metrics/active-contribution-weeks/helpers/get-weeks-of-created-repositories.ts
@@ -1,9 +1,6 @@
-import {
-  WeekObject,
-  CreatedRepositories,
-} from '@/github/metrics/active-contribution-weeks/types';
-import { getOpenSourceRepositories } from '@/github/metrics/helpers';
+import { getOpenSourceContributionsRepositories } from '@/github/metrics/helpers';
 import { transformToWeekObject } from './transform-to-week-object';
+import { WeekObject, CreatedRepositories } from '../types';
 
 /**
  * Collect all occurredAt and convert to week object for created repositories
@@ -12,7 +9,8 @@ export function getWeeksOfCreatedRepositories(
   createdRepositories: CreatedRepositories[],
 ): WeekObject[] {
   // Get all contributions from open source repositories
-  const openSourceRepositories = getOpenSourceRepositories(createdRepositories);
+  const openSourceRepositories =
+    getOpenSourceContributionsRepositories(createdRepositories);
   // Collect and transform weeks
   const weeks: WeekObject[] = [];
   for (const repository of openSourceRepositories) {
