+++ src/github/metrics/active-contribution-weeks/helpers/index.ts
@@ -1,3 +1,4 @@
 export * from './get-weeks-of-contributions';
+export * from './get-weeks-of-created-repositories';
 export * from './remove-duplicate-weeks';
 export * from './transform-to-week-object';
