+++ src/github/metrics/helpers/get-open-source-repositories.ts
@@ -1,11 +1,11 @@
 const openSourceLicenses = ['mit', 'apache-2.0', 'gpl-3.0', 'agpl-3.0'];
 
 /**
- * Return open source repositories.
+ * Return open source contributions repositories.
  */
-export function getOpenSourceRepositories<T extends RepositoriesQuery>(
-  contributionsByRepositories: T[],
-): T[] {
+export function getOpenSourceContributionsRepositories<
+  T extends RepositoriesQuery,
+>(contributionsByRepositories: T[]): T[] {
   return contributionsByRepositories.filter(
     (contributionsByRepository) =>
       !contributionsByRepository.repository.isFork &&
@@ -16,14 +16,28 @@ export function getOpenSourceRepositories<T extends RepositoriesQuery>(
   );
 }
 
+/**
+ * Return open source repositories.
+ */
+export function getOpenSourceRepositories<T extends RepositoryQuery>(
+  repositories: T[],
+): T[] {
+  return repositories.filter(
+    (repository) =>
+      !repository.isFork &&
+      !repository.isPrivate &&
+      openSourceLicenses.includes(repository.licenseInfo?.key),
+  );
+}
+
 export interface RepositoriesQuery {
   repository: RepositoryQuery;
 }
 
-interface RepositoryQuery {
+export interface RepositoryQuery {
   licenseInfo?: LicenseQuery;
   isFork: boolean;
-  isPrivate: boolean;
+  isPrivate?: boolean;
 }
 
 interface LicenseQuery {
