+++ src/github/metrics/index.ts
@@ -1,9 +1,9 @@
 export * from './active-contribution-weeks/active-contribution-weeks.metric';
+export * from './contributed-repository-count/contributed-repository-count.metric';
+export * from './issue-pull-request-ratio/issue-pull-request-ratio.metric';
+export * from './maintained-repository-count/maintained-repository-count.metric';
+export * from './total-contributions/total-contributions.metric';
+export * from './user-info/user-info.metric';
 export * from './base.metric';
-export * from './contributed-repository-count.metric';
-export * from './issue-pull-request-ratio.metric';
-export * from './maintained-repository-count.metric';
 export * from './metric-name';
 export * from './metricable';
-export * from './total-contributions.metric';
-export * from './user-info.metric';
