+++ src/github/metrics/issue-pull-request-ratio/helpers/get-contributions-count.ts
@@ -1,0 +1,20 @@
+import { getOpenSourceContributionsRepositories } from '@/github/metrics/helpers';
+import { ContributionsByRepository } from '../types';
+
+/**
+ * Return contributions count
+ */
+export function getContributionsCount(
+  contributionsByRepositories: ContributionsByRepository[],
+): number {
+  // Get all contributions from open source repositories
+  const openSourceContributionsByRepositories =
+    getOpenSourceContributionsRepositories(contributionsByRepositories);
+  // Get total of contributions
+  let contributionsCount = 0;
+  openSourceContributionsByRepositories.forEach((contributionByRepository) => {
+    contributionsCount =
+      contributionsCount + contributionByRepository.contributions.totalCount;
+  });
+  return contributionsCount;
+}
