+++ src/github/metrics/issue-pull-request-ratio/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './get-contributions-count';
