+++ src/github/metrics/issue-pull-request-ratio/types/contributions-by-repositories.interface.ts
@@ -1,0 +1,7 @@
+import { RepositoriesQuery } from '@/github/metrics/helpers';
+
+export interface ContributionsByRepository extends RepositoriesQuery {
+  contributions: {
+    totalCount: number;
+  };
+}
