+++ src/github/metrics/issue-pull-request-ratio/types/index.ts
@@ -1,0 +1,1 @@
+export * from './contributions-by-repositories.interface';
