+++ src/github/metrics/total-contributions/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './get-contributions-count';
