+++ src/github/metrics/total-contributions/types/index.ts
@@ -1,0 +1,1 @@
+export * from './contributions-by-repositories.interface';
